tratamento de excecao - delete

+++ src/main/java/com/dhoinformatica/curso/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.dhoinformatica.curso.entities.User;
 import com.dhoinformatica.curso.repositories.UserRepository;
+import com.dhoinformatica.curso.resources.exceptions.DatabaseException;
 import com.dhoinformatica.curso.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -29,8 +31,17 @@ public class UserService {
 		return repository.save(obj);
 	}
 	
-	public void delete(Long id) {
-		repository.deleteById(id);
+	public void delete(Long id){
+	    try {
+	        if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+	        repository.deleteById(id);
+	    }
+	    catch (ResourceNotFoundException e){
+	        throw new ResourceNotFoundException(id);
+	    }
+	    catch(DataIntegrityViolationException e) {
+	    	throw new DatabaseException(e.getMessage());
+	    }
 	}
 	
 	public User update(Long id, User obj) {
